+++ src/main/java/com/diego/gestorcasino/models/Plato.java
@@ -8,7 +8,7 @@ public class Plato {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column(nullable = false)
     private String nombre;
@@ -23,11 +23,13 @@ public class Plato {
     private String categoria;
 
     // Getters y setters
-    public Long getId() {
+
+
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
